+++ app/src/main/java/com/gabrielMJr/twaire/physic/units/AccelerationIF.java
@@ -1,12 +1,14 @@
 package com.gabrielMJr.twaire.physic.units;
 
+import java.math.BigDecimal;
+
 public interface AccelerationIF 
 {
-    public abstract Double toKilometerSquarePerSecond(double value, String unit);
-	public abstract Double toHectometerSquarePerSecond(double value, String unit);
-	public abstract Double toDecameterSquarePerSecond(double value, String unit);
-	public abstract Double toMeterSquarePerSecond(double value, String unit);
-	public abstract Double toDecimeterSquarePerSecond(double value, String unit);
-	public abstract Double toCentimeterSquarePerSecond(double value, String unit);
-	public abstract Double toMillimeterSquarePerSecond(double value, String unit);
+    public abstract BigDecimal toKilometerSquarePerSecond(double value, String unit);
+	public abstract BigDecimal toHectometerSquarePerSecond(double value, String unit);
+	public abstract BigDecimal toDecameterSquarePerSecond(double value, String unit);
+	public abstract BigDecimal toMeterSquarePerSecond(double value, String unit);
+	public abstract BigDecimal toDecimeterSquarePerSecond(double value, String unit);
+	public abstract BigDecimal toCentimeterSquarePerSecond(double value, String unit);
+	public abstract BigDecimal toMillimeterSquarePerSecond(double value, String unit);
 }
